- Moves commonly used libs into lib/common - deletes unused download-lib
