+++ jobs/atc.js
@@ -4,11 +4,11 @@ var path = require("path");
 
 var cfg = require("./cfg/atc.cfg.js");
 var log = require("../lib").log;
-var disk = require("../lib/disk");
-var fetchHTML = require("../lib/fetchHTML");
-var parseLink = require("../lib/parseLink");
-var downloadFile = require("../lib/downloadFile");
-var renderProgress = require("../lib/renderProgress");
+var disk = require("../lib/common/disk");
+var fetchHTML = require("../lib/common/fetchHTML");
+var parseLink = require("../lib/common/parseLink");
+var downloadFile = require("../lib/common/downloadFile");
+var renderProgress = require("../lib/common/renderProgress");
 var readXLSX = require("../lib/atc/readXLSX");
 var addCodes = require("../lib/atc/addCodes");
 var modifyCodes = require("../lib/atc/modifyCodes");
