+++ lib/atc/getSwissmedicData.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var disk = require("../disk");
+var disk = require("../common/disk");
 var swissmedicJob = require("../../jobs/swissmedic");
 var swissmedicCfg = swissmedicJob.cfg;
