+++ lib/atc/modifyNames.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var disk = require('../disk');
+var disk = require('../common/disk');
 
 /**
  *
