+++ lib/bag/parseBAGXML.js
@@ -4,7 +4,7 @@ var fs = require("fs");
 
 var splitter = require('xml-splitter');
 
-var util = require("../util");
+var util = require("../common/util");
 var log = require("../").log;
 
 /**
