+++ lib/common/disk.js
