+++ lib/common/parseLink.js
