+++ lib/common/renderProgress.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var log = require("./").log;
+var log = require("./../index").log;
 
 /**
  *
