+++ lib/common/util.js
