+++ lib/index.js
@@ -6,8 +6,7 @@ module.exports.log = log({
 });
 
 exports.emil = require("./emil.js");
-exports.util = require("./util.js");
-exports.download = require("./download.js");
+exports.util = require("./common/util.js");
 
 exports.atc = require("../jobs/atc");
 exports.bag = require("../jobs/bag");
