+++ lib/kompendium/startDownload.js
@@ -3,7 +3,7 @@
 var jsdom = require("jsdom").jsdom;
 
 var log = require("../").log;
-var disk = require("../disk");
+var disk = require("../common/disk");
 
 /**
  *
