+++ lib/swissmedic/createATCCorrection.js
@@ -13,7 +13,7 @@
  * @property {ATCData} id
  */
 
-var disk = require("../../lib/disk");
+var disk = require("../common/disk");
 
 /**
  *
