+++ lib/swissmedic/readXLSX.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var xlsx = require("xlsx");
-var util = require("../util");
+var util = require("../common/util");
 
 /**
  *
