+++ test/unit/lib/common/parseLink.test.js
@@ -11,14 +11,14 @@ describe("parseLinks", function () {
   var parseLink, url, html, regex, ref, jobCfg;
 
   before(function () {
-    parseLink = require("../../../lib/parseLink");
+    parseLink = require("../../../../lib/common/parseLink");
   });
 
   describe("swissmedic", function () {
     before(function () {
-      jobCfg = require("../../../jobs/cfg/swissmedic.cfg");
+      jobCfg = require("../../../../jobs/cfg/swissmedic.cfg.js");
       url = "https://www.swissmedic.ch/arzneimittel/00156/00221/00222/00230/index.html";
-      html = fs.readFileSync(path.resolve(__dirname, "../../fixtures/html/swissmedic.html"), {encoding: "utf8"});
+      html = fs.readFileSync(path.resolve(__dirname, "../../../fixtures/html/swissmedic.html"), {encoding: "utf8"});
       ref = url + "?lang=de&download=NHzLpZeg7t,lnp6I0NTU042l2Z6ln1acy4Zn4Z2qZpnO2Yuq2Z6gpJCDdHx7hGym162epYbg2c_JjKbNoKSn6A--";
     });
 
@@ -39,9 +39,9 @@ describe("parseLinks", function () {
 
   describe("atc", function () {
     before(function () {
-      jobCfg = require("../../../jobs/cfg/atc.cfg");
+      jobCfg = require("../../../../jobs/cfg/atc.cfg.js");
       url = "http://www.wido.de/amtl_atc-code.html";
-      html = fs.readFileSync(path.resolve(__dirname, "../../fixtures/html/atc.html"), {encoding: "utf8"});
+      html = fs.readFileSync(path.resolve(__dirname, "../../../fixtures/html/atc.html"), {encoding: "utf8"});
       ref = "http://www.wido.de/fileadmin/wido/downloads/pdf_arzneimittel/atc/wido_arz_amtl_atc-index_1214.zip";
     });
 
@@ -57,9 +57,9 @@ describe("parseLinks", function () {
 
   describe("bag", function () {
     before(function () {
-      jobCfg = require("../../../jobs/cfg/bag.cfg");
+      jobCfg = require("../../../../jobs/cfg/bag.cfg.js");
       url = "http://www.spezialitaetenliste.ch/";
-      html = fs.readFileSync(path.resolve(__dirname, "../../fixtures/html/bag.html"), {encoding: "utf8"});
+      html = fs.readFileSync(path.resolve(__dirname, "../../../fixtures/html/bag.html"), {encoding: "utf8"});
       ref =  "http://www.spezialitaetenliste.ch/File.axd?file=XMLPublications.zip";
     });
